+++ src/main/java/com/codecrafters/msadminusers/service/UserService.java
@@ -10,7 +10,7 @@ public interface UserService {
 
     Account getUserById(Integer id);
 
-    void deleteById(Long id);
+    void deleteById(Integer id);
 
     Account save(Account user);
